split ImageCopier
